voice-agents/background_agent/pyproject.toml
[project]
name = "atoms-background-agent"
version = "0.1.0"
description = "Background processing with multi-node agent architecture"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "smallestai>=4.2.0",
    "python-dotenv",
    "loguru",
]
